Create README.md
This is a simple web API "Hello World". It outputs "Hello World" to the console and also returns that string in an HTTP response.

The API is separated from the program logic. As a web API, it could service many different types of applications.

Inheritance is utilized to account for future enhancements for writing to different types of destination (database, console app, etc).

A configuration file is used to determine where it will write/output.

A simple test is used to verify the output, but many more could be added, especially as features are added.